Rebase ExaConstit so it is now a velocity-based FEM code
Rewrote ExaConstit so that it solves for the velocity field instead of the displacement field. The advantages of this are that many crystal plasticity methods use quantities derived from the velocity gradient. Therefore, it's a lot easier to deal with these quantities when your solution field is the velocity field. One can also think of this as changing the code base from an Updated Total Lagrangian formulation to an Updated Lagrangian formulation.