(Claude) Beef up the UMAT handling / dload aspect of things Since spent enough time getting the dynamic umat loader stuff working decided to just port that over to our internal UMAT stack as that way I don't have to keep trying to figure out compiler name mangling when moving between compilers... Now at runtime it figures it out for us on the first call and then saves that for future calls. I had Claude help with this aspect of things. It also includes some debugging aspect to make it easy to figure out where something is coming from. Then if it can't find a function call as there might be a mistake in the user provided options it'll search the library for a number of different typical umat function names and if it finds one it'll use it. Although, we do warn the user about this.
I've also been working on a better UMAT wrapper example and will likely upload that at some point for users in the near term as it also just finds the UMAT func name for them and makes things simpler not harder.
Outside of that there were some minor issues with the dynamic library accidentally not finding the library paths that we wanted and ultimately calling our internal umat as it was ignoring the path we provided.