Vast formatting changes for consistency everywhere more or less in at least fcn capitalization For anything that is not IO related (anything but src/options and src/utilities/unified_logger.*) all the function signatures have been converted to "PascalCase" in-order to be consistent with how MFEM does things and just standardize things.
IO related files (src/options/* and src/utilities/unified_logger.*) all of the functions were set to "snake_case" as this was cleaner looking for the options work and then for the loggers it better matched what the STDLIB is doing which makes things easier to read / reason about. I honestly would prefer all functions signatures look like this but with how MFEM does things that just clashed too much everywhere else :(

Now there were also some member variable name changes any member variable that ended with the postfix "_" had that postfix removed. This addition is something that Claude just loves to add... and I finally got a round to removing them. I still need to make the member variable names consistent everywhere but that will likely be a fight for another day...

Note, I did try using clang-tidy to enforce this automatically but it kept getting too many things wrong or changing too much. So, I just abandonded that idea of using it as a tool to automatically do this for us. I might still look into using it for other things though as well as leveraging like clang-format.